fix: upd.  app.js
