+++ src/utils/utils.js
@@ -3,3 +3,11 @@ export const toCapitalizeText = (text) => {
 }
 export const greeting  = (username) => console.log(`Welcome to the File Manager, ${toCapitalizeText(username)} !`);
 export const goodbye  = (username) => console.log(`Thank you for using File Manager, ${toCapitalizeText(username)} goodbye!`);
+
+export const tryCatchWrapper = async (fn, arg) => {
+      try {
+        return await fn(arg);
+      } catch(e) {
+        console.log('Operation failed');
+      }
+}
